requirements.txt
pyzbar
opencv-python
psycopg2-binary
serial
pyserial